Extract interactive command handling into InteractiveSession
